Add files via upload
Added movetostartofline.lua, a plugin that hijacks the home button; when you press home it will take you to the start of the line, but exluding any starting whitespace of that line. If You already are at the first non-whitespace character of that line, it will move the cursor to the start of the line, including whitespace characters.